# money-exchange/src/main/java/pe/com/nttdata/moneyexchange/dto/DivisaDto.java
package pe.com.nttdata.moneyexchange.dto;

import lombok.Getter;
import lombok.Setter;

import java.time.LocalDateTime;

@Getter
@Setter
public class DivisaDto {

    private double buy;
    private double sell;
    private String date;

    public double getBuy() {
        return buy;
    }

    public void setBuy(double buy) {
        this.buy = buy;
    }

    public double getSell() {
        return sell;
    }

    public void setSell(double sell) {
        this.sell = sell;
    }

    public String getDate() {
        return date;
    }

    public void setDate(String date) {
        this.date = date;
    }
}